Introduce client-side baskets.

+++ plugins/Baskets/src/EasyAbp.EShop.Plugins.Baskets.Web/Pages/EShop/Plugins/Baskets/BasketItems/BasketItem/index.js
@@ -1,22 +1,23 @@
 $(function () {
 
+    var localStorageItemKey = "EShopBasket:" + basketName;
+
     var l = abp.localization.getResource('EasyAbpEShopPluginsBaskets');
+    
+    var serverSide = abp.setting.getBoolean('EasyAbp.EShop.Plugins.Baskets.EnableServerSideBaskets')
+        && abp.auth.isGranted('EasyAbp.EShop.Plugins.Baskets.BasketItem');
 
     var service = easyAbp.eShop.plugins.baskets.basketItems.basketItem;
     var createModal = new abp.ModalManager(abp.appPath + 'EShop/Plugins/Baskets/BasketItems/BasketItem/CreateModal');
     var editModal = new abp.ModalManager(abp.appPath + 'EShop/Plugins/Baskets/BasketItems/BasketItem/EditModal');
 
-    var dataTable = $('#BasketItemTable').DataTable(abp.libs.datatables.normalizeConfiguration({
+    var configuration = {
         processing: true,
-        serverSide: true,
         paging: true,
         searching: false,
         autoWidth: false,
         scrollCollapse: true,
         order: [[1, "asc"]],
-        ajax: abp.libs.datatables.createAjax(service.getList, function () {
-            return { basketName: basketName, userId: userId }
-        }),
         columnDefs: [
             {
                 rowAction: {
@@ -24,30 +25,34 @@ $(function () {
                         [
                             {
                                 text: l('Edit'),
-                                visible: abp.auth.isGranted('EasyAbp.EShop.Plugins.Baskets.BasketItem.Update'),
                                 action: function (data) {
-                                    editModal.open({ id: data.record.id });
+                                    editModal.open({ basketName: basketName, id: data.record.id });
                                 }
                             },
                             {
                                 text: l('Delete'),
-                                visible: abp.auth.isGranted('EasyAbp.EShop.Plugins.Baskets.BasketItem.Delete'),
                                 confirmMessage: function (data) {
                                     return l('BasketItemDeletionConfirmationMessage', data.record.id);
                                 },
                                 action: function (data) {
+                                    if (serverSide) {
                                         service.delete(data.record.id)
-                                        .then(function () {
-                                            abp.notify.info(l('SuccessfullyDeleted'));
-                                            dataTable.ajax.reload();
-                                        });
+                                            .then(function () {
+                                                abp.notify.info(l('SuccessfullyDeleted'));
+                                                dataTable.ajax.reload();
+                                            });
+                                    } else {
+                                        var cachedItems = JSON.parse(localStorage.getItem(localStorageItemKey)) || [];
+                                        cachedItems.splice(cachedItems.findIndex(x => x.id === data.record.id), 1);
+                                        localStorage.setItem(localStorageItemKey, JSON.stringify(cachedItems));
+                                        location.reload();
+                                    }
                                 }
                             }
                         ]
                 }
             },
             { data: "basketName" },
-            { data: "userId" },
             { data: "storeId" },
             { data: "productId" },
             { data: "productSkuId" },
@@ -64,15 +69,67 @@ $(function () {
             { data: "inventory" },
             { data: "isInvalid" },
         ]
-    }));
+    };
+
+    var dataTable;
+    
+    if (serverSide) {
+        configuration.serverSide = true;
+        configuration.ajax = abp.libs.datatables.createAjax(service.getList, function () {
+            return { basketName: basketName, userId: userId }
+        });
+        
+        var clientSideItems = JSON.parse(localStorage.getItem(localStorageItemKey)) || [];
+
+        // Move client-side basket items to server-side.
+        if (clientSideItems.length > 0) {
+            localStorage.setItem(localStorageItemKey, JSON.stringify([]));
+            createManyServerSideBasketItems(clientSideItems);
+        }
+
+        dataTable = $('#BasketItemTable').DataTable(abp.libs.datatables.normalizeConfiguration(configuration));
+    } else {
+        configuration.serverSide = false;
+        var cachedItems = JSON.parse(localStorage.getItem(localStorageItemKey)) || [];
+        if (cachedItems.length > 0) {
+            service.generateClientSideData({ items: cachedItems }).then(function (result) {
+                configuration.data = result.items
+                dataTable = $('#BasketItemTable').DataTable(abp.libs.datatables.normalizeConfiguration(configuration));
+            });
+        } else {
+            configuration.data = [];
+            dataTable = $('#BasketItemTable').DataTable(abp.libs.datatables.normalizeConfiguration(configuration));
+        }
+    }
 
     createModal.onResult(function () {
-        dataTable.ajax.reload();
+        if (serverSide) {
+            dataTable.ajax.reload();
+        } else {
+            location.reload();
+        }
     });
 
     editModal.onResult(function () {
-        dataTable.ajax.reload();
+        if (serverSide) {
+            dataTable.ajax.reload();
+        } else {
+            location.reload();
+        }
     });
+    
+    function createManyServerSideBasketItems(items, autoReloadDataTable = true) {
+        var item = items.shift();
+        service.create(item, {
+            success: function () {
+                if (items.length > 0) {
+                    createManyServerSideBasketItems(items);
+                } else if (autoReloadDataTable) {
+                    dataTable.ajax.reload();
+                }
+            }
+        })
+    }
 
     $('#NewBasketItemButton').click(function (e) {
         e.preventDefault();
